ml_service/requirements.txt
flask
psycopg2-binary
joblib
scikit-learn==1.8.0
apscheduler
requests
pandas